+++ src/controller/include/safetyboard.h
@@ -24,6 +24,9 @@ extern "C" {
 
 //functions
 //uint8_t crc8(uint8_t* message, size_t len, uint8_t poly, uint8_t init);
+void crcInit(void);
+uint8_t crc8Fast(uint8_t const *message, int len);
+
 uint8_t crc8(uint8_t* message, int len);
 //uint8_t crc8(unsigned char* message, int len);
 uint8_t paddedCRC(uint8_t* message, size_t len);
@@ -32,6 +35,10 @@ bool checkCRC(uint8_t CRC, uint8_t* payload, int len);
 uint8_t extractCRC(uint8_t* payload, int len);
 
 bool checkRC(uint8_t currRC, uint8_t prevRC, uint8_t RCMax);
+//bool checkRC(uint8_t currRC, uint8_t prevRC);
+void updateRC(void);
+uint8_t getRollingCount(void);
+
 bool checkCommandRange(uint8_t actReq);
 uint8_t saturateOutputCommand(uint8_t actReq, uint8_t actReqPrev);
 bool determineState(bool crcFaultActive, bool rcFaultActive, bool wheelSpeedFaultActive, bool outOFRangeFaultActive);
